refactor: follow 서비스 및 레포지토리 테스트 수정
