+++ backend/src/main/java/org/cosmic/backend/domain/user/repositorys/FollowRepository.java
@@ -1,6 +1,7 @@
 package org.cosmic.backend.domain.user.repositorys;
 
 import java.util.List;
+import java.util.Optional;
 import org.cosmic.backend.domain.user.domains.Follow;
 import org.cosmic.backend.domain.user.domains.FollowPK;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -12,4 +13,6 @@ public interface FollowRepository extends JpaRepository<Follow, FollowPK> {
   List<Follow> findAllByUser_UserId(Long userId);
 
   List<Follow> findAllByOther_UserId(Long userId);
+
+  Optional<Follow> findByUser_UserIdAndOther_UserId(Long userId, Long otherUserId);
 }
